refactor(map): extract shared IpBanControls out of MapPopup
