+++ resources/components/map/LiveRequestPopup.tsx
@@ -3,10 +3,9 @@
  * since MapLibre popups render outside the Tailwind-scoped tree.
  */
 import { Popup } from "react-map-gl/maplibre"
-import { Loader2, ShieldBan, ShieldOff } from "lucide-react"
 import { formatBytes, formatDuration } from "@/lib/api"
-import { useBanIp, useBannedIps, useCrowdsecStatus, useUnbanIp } from "@/lib/queries"
 import { PACKET_COLORS } from "@/lib/live-traffic/classify"
+import { IpBanControls } from "./IpBanControls"
 import type { LiveRequest } from "@/lib/live-traffic/types"
 
 function Row({ label, value }: { label: string; value: string }) {
@@ -25,12 +24,6 @@ export function LiveRequestPopup({
   request: LiveRequest
   onClose: () => void
 }) {
-  const { data: status } = useCrowdsecStatus()
-  const { data: bannedIps } = useBannedIps()
-  const ban = useBanIp()
-  const unban = useUnbanIp()
-  const banned = bannedIps?.has(request.ip) ?? request.banned
-  const isPending = ban.isPending || unban.isPending
   const log = request.log
   const [longitude, latitude] = request.coordinates ?? [0, 0]
 
@@ -107,44 +100,7 @@ export function LiveRequestPopup({
         {log?.referrer && <Row label="Referrer" value={log.referrer} />}
         {log?.user_agent && <Row label="Agent" value={log.user_agent} />}
 
-        {(banned || status?.write_enabled) && (
-          <div style={{ display: "flex", alignItems: "center", gap: "8px", marginTop: "8px", paddingTop: "8px", borderTop: "1px solid var(--popup-border)" }}>
-            {banned && (
-              <span style={{ fontSize: "9px", fontWeight: 600, textTransform: "uppercase", color: "#f87171", background: "rgba(239, 68, 68, 0.15)", padding: "1px 6px", borderRadius: "9999px" }}>
-                banned
-              </span>
-            )}
-            {status?.write_enabled && (
-              <button
-                onClick={() => (banned ? unban.mutate(request.ip) : ban.mutate({ ip: request.ip }))}
-                disabled={isPending}
-                aria-label={banned ? `Unban ${request.ip}` : `Ban ${request.ip}`}
-                style={{
-                  marginLeft: "auto",
-                  display: "inline-flex",
-                  alignItems: "center",
-                  gap: "4px",
-                  background: "transparent",
-                  border: "1px solid var(--popup-border)",
-                  borderRadius: "4px",
-                  padding: "2px 8px",
-                  fontSize: "10px",
-                  cursor: isPending ? "wait" : "pointer",
-                  color: "var(--popup-muted)",
-                }}
-              >
-                {isPending ? (
-                  <Loader2 style={{ width: 12, height: 12, animation: "spin 1s linear infinite" }} />
-                ) : banned ? (
-                  <ShieldOff style={{ width: 12, height: 12 }} />
-                ) : (
-                  <ShieldBan style={{ width: 12, height: 12 }} />
-                )}
-                {banned ? "Unban" : "Ban"}
-              </button>
-            )}
-          </div>
-        )}
+        <IpBanControls ip={request.ip} initialBanned={request.banned} variant="footer" />
       </div>
     </Popup>
   )
